Add CompositeParamDeepCopyUtil
