+++ dubbo-rpc/dubbo-rpc-triple/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.protocol.injvm.ParamDeepCopyUtil
@@ -1,0 +1,1 @@
+composite=org.apache.dubbo.rpc.protocol.injvm.CompositeParamDeepCopyUtil
